// src/app/models/ticket.model.ts
export class Ticket {
    ticketId?: number;
    description?: string;
    shortDescription?: string;
    category?: string;
    subCategory?: string; 
    ticketState?: string;
    ticketUrgency?: string;
    openedBy?: string;
    assignedBy?: string;
    assignedTo?: string;
    assignedTeam?: string;
    openedDate?: Date;
    resolutionComment?: string;
    resolutionDate?: Date;
    workEmail?: string;
  }
  